+++ .github/workflows/try.yml
@@ -1,10 +1,10 @@
 name: Module Tests
 on:
   push:
-    branches: main
+    branches: master
     paths: '**.lua'
   pull_request:
-    branches: main
+    branches: master
     paths: '**.lua'
 jobs:
   Vanilla:
