+++ tests/common-spec/bit.lua
