+++ tests/common-spec/byte.lua
