+++ tests/common-spec/int16.lua
