+++ tests/common-spec/int8.lua
