+++ tests/common-spec/setlengthstring.lua
