+++ tests/common-spec/uint32.lua
