+++ tests/roblox-spec/colorsequence.lua
